add separate component Restaurant
